Image-Watermarking-App-main/requirements.txt
Pillow==10.1.0
tkinterdnd2_universal==1.7.3
